+++ STm32F4_Firmware/src/fuel.h
@@ -17,8 +17,8 @@ public:
 	float get_fuel(char char_buffer[]);
 	void init();
 	int check_vars();
-	int blink_start; // bei mehr als 140 km auf der Tankuhr fngt die an zu blinken
-	int blink_freq; // blinkt mit 500ms Periodenlaenge
+	int16_t blink_start; // bei mehr als 140 km auf der Tankuhr fngt die an zu blinken
+	int16_t blink_freq; // blinkt mit 500ms Periodenlaenge
 };
 
 #endif /* FUEL_H_ */
